% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostparasite.R
\docType{data}
\name{hostparasite}
\alias{hostparasite}
\title{A binarized host-parasite network}
\format{
A matrix of binary entries with 27 rows and 26 columns
}
\source{
\url{https://www.web-of-life.es/}

\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.jf3tj}
}
\usage{
hostparasite
}
\description{
Locality: Tomsk-Tumen
Reference:Hadfield JD, Krasnov BR, Poulin R, Shinichi N (2013) A tale of two phylogenies: comparative analyses of ecological interactions. The American Naturalist 183(2): 174-187
}
\keyword{datasets}
