% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_emp.R
\name{plot.robber}
\alias{plot.robber}
\title{Plot function pour robber class}
\usage{
\method{plot}{robber}(x, type = "function", add = FALSE, ...)
}
\arguments{
\item{x}{An object of class robber}

\item{type}{Type of the plot, just function}

\item{add}{Whether it should be added to a previous ggplot}

\item{...}{\code{lty}, \code{col}, ... to be passed to \code{geom_step()} in
\code{ggplot2}}
}
\value{
A \code{ggplot2} object
}
\description{
Plot function pour robber class
}
\examples{
data("hostparasite", package = "robber")
rob <- robustness_emp(hostparasite)
plot(rob)
}
