% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollination.R
\docType{data}
\name{pollination}
\alias{pollination}
\title{A plant-pollinator network}
\format{
A matrix of binary entries with 17 rows and 51 columns
}
\source{
\url{https://www.web-of-life.es/}
}
\usage{
pollination
}
\description{
Locality: Tenerife, Canary Islands
Stald L (2003). Struktur og dynamik i rum og tid af et bestøvningsnetværk på Tenerife, De Kanariske Øer. Msc thesis (Univ of Aarhus, Aarhus, Denmark)
}
\keyword{datasets}
