% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_emp.R
\name{robustness_emp}
\alias{robustness_emp}
\title{Empirical Robustness of a Network}
\usage{
robustness_emp(A, ext_seq = "uniform", nb_iter = 300, net = "bipartite")
}
\arguments{
\item{A}{A binary incident matrix}

\item{ext_seq}{A string, the rule for the primary extinctions sequences,
one of \code{"uniform"}, the default for uniform extinctions sequences,
\code{"decreasing"} and \code{"increasing"} for primary extinctions sequences
by increasing and decreasing degree order sequence on the row species}

\item{nb_iter}{An integer, the number of generated sequences. Default to
\code{300}}

\item{net}{A string, the type of network. Only "bipartite".}
}
\value{
A list of size 2:
\itemize{
\item \code{$fun} is the robustness function, a vector of size \code{(nrow(A) +1)}
\item \code{$auc} the area under the curve of the robustness function
}
}
\description{
Compute the robustness of an ecological network by averaging
over a great number of randomly generated primary extinctions sequences
}
\examples{
A <- matrix(c(1, 0, 0,
             0, 1, 0,
             0, 0, 1,
             1, 1, 1), nrow = 4, ncol = 3, byrow = TRUE)
my_rob <- robustness_emp(A, ext_seq = "increasing")
my_rob$fun
my_rob$auc
}
