% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeddispersal.R
\docType{data}
\name{seeddispersal}
\alias{seeddispersal}
\title{A seed-dispersal network}
\format{
A matrix of binary entries with 36 rows and 19 columns
}
\source{
\url{http://www.web-of-life.es/}

\url{https://conbio.org/publications/conservation-biology}
}
\usage{
seeddispersal
}
\description{
Locality: North Negros Forest Reserve, Central Philippi
Hammann, A. & Curio, B. (1999) Interactions among frugivores and fleshy fruit trees in a Philippine submontane rainforest.
}
\keyword{datasets}
