% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{var_auc_unif_lbm_cpp}
\alias{var_auc_unif_lbm_cpp}
\title{Compute the Conditional Variance of the Robustness}
\usage{
var_auc_unif_lbm_cpp(con, pi, rho, nr, nc)
}
\arguments{
\item{con}{A matrix, the connectivity parameter}

\item{pi}{A vector of length \code{nrow(con)}, the proportion of row blocks}

\item{rho}{A vector of length \code{ncol(con)}, the proportion of
column blocks}

\item{nr}{An integer, the number of row (primary) species}

\item{nc}{An integer, the number of column (secondary) species}
}
\description{
Compute Var_A(E_S\link{AUC | A})
}
