% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_bm.R
\name{robustness_lbm}
\alias{robustness_lbm}
\title{Robustness for LBM}
\usage{
robustness_lbm(con, pi, rho, nr, nc, ext_seq = "uniform")
}
\arguments{
\item{con}{A matrix, the connectivity parameter}

\item{pi}{A vector of length \code{nrow(con)}, the proportion of row blocks}

\item{rho}{A vector of length \code{ncol(con)}, the proportion of
column blocks}

\item{nr}{An integer, the number of row (primary) species}

\item{nc}{An integer, the number of column (secondary) species}

\item{ext_seq}{A string, the rule for the primary extinctions sequences,
one of \code{"uniform"}, the default for uniform extinctions sequences,
\code{"decreasing"} and \code{"increasing"} for primary extinctions sequences
by increasing and decreasing row blocks connection probability}
}
\value{
A list of size 2:
\itemize{
\item \code{$fun} is the robustness function, a vector of size \code{nr +1}
\item \code{$auc} the area under the curve of the robustness function
}
}
\description{
Compute the robustness for a given set of Latent Block Model
parameters.
}
\examples{
con <- matrix(c(.5,.3,.3,.1), 2, 2)
pi  <- c(.25,.75)
rho <- c(1/3, 2/3)
nr <- 50
nc <- 30
my_rob <- robustness_lbm(con, pi, rho, nr, nc)
my_rob$fun
my_rob$auc
}
